TP WEBSERVICE (WORK IN PROGRESS) : CRUD MVC + VIEWS  pour toutes les ressources + Controller Rest pour toutes les ressources

+++ src/main/java/com/bigcorp/booking/dao/spring/ArticleSpringDao.java
@@ -1,17 +1,16 @@
 package com.bigcorp.booking.dao.spring;
 
 import java.util.List;
-import java.util.Optional;
+
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import com.bigcorp.booking.model.Article;
-import com.bigcorp.booking.model.Fournisseur;
+
 
 @Repository
-public interface ArticleSpringDao extends CrudRepository<Article, Long> {
+public interface ArticleSpringDao extends CrudRepository<Article, Integer> {
 	 
 	List<Article>findByNom(String name);
-	
 }
