+++ src/main/java/com/bigcorp/booking/dao/spring/ClientSpringDao.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import com.bigcorp.booking.model.Client;
 
 @Repository
-public interface ClientSpringDao  extends CrudRepository<Client, Long>{
+public interface ClientSpringDao  extends CrudRepository<Client, Integer>{
 	List<Client> findByNom(String nom);
 }
